requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
mongoengine>=0.24.0
requests
pillow